identification categories works with both index styles; extract some observation preloading into new model
