+++ lib/controllers/v1/observations_controller.js
@@ -2089,15 +2089,30 @@ ObservationsController.subscriptions = ( req, callback ) => {
 ObservationsController.identificationCategories = ( req, callback ) => {
   const countQuery = _.assignIn( { }, req.query );
   countQuery.aggs = {
+    nested: {
+      nested: { path: "identifications" },
+      aggs: {
+        categories: { terms: { field: "identifications.category", size: 10 } }
+      }
+    },
     categories: { terms: { field: "identification_categories", size: 10 } }
   };
   countQuery.per_page = 0;
   const countReq = _.assignIn( { }, req, { query: countQuery } );
   ObservationsController.elasticResults( countReq, ( err, data ) => {
     if ( err ) { return void callback( err ); }
-    const buckets = _.map( data.aggregations.categories.buckets, b => (
-      { category: b.key, count: b.doc_count }
-    ) );
+    const combinedCategories = { };
+    _.each( data.aggregations.nested.categories.buckets, b => {
+      combinedCategories[b.key] = combinedCategories[b.key] || 0;
+      combinedCategories[b.key] += b.doc_count;
+    } );
+    _.each( data.aggregations.categories.buckets, b => {
+      combinedCategories[b.key] = combinedCategories[b.key] || 0;
+      combinedCategories[b.key] += b.doc_count;
+    } );
+    const buckets = _.map( combinedCategories, ( v, k ) => ( {
+      category: k, count: v
+    } ) );
     callback( null, {
       total_results: _.size( buckets ),
       page: 1,
