+++ lib/models/observation.js
@@ -1,30 +1,15 @@
 const _ = require( "lodash" );
-const squel = require( "squel" );
-const URL = require( "url" );
 const util = require( "../util" );
 const ControlledTerm = require( "./controlled_term" );
 const ObservationField = require( "./observation_field" );
 const ESModel = require( "./es_model" );
-const pgClient = require( "../pg_client" );
 const Identification = require( "./identification" );
 const Project = require( "./project" );
 const Taxon = require( "./taxon" );
 const User = require( "./user" );
 const Model = require( "./model" );
 const TaxaController = require( "../controllers/v1/taxa_controller" );
-
-const LICENSE_INFO = {
-  0: { code: "c", short: "(c)" },
-  1: { code: "cc-by-nc-sa", short: "CC BY-NC-SA" },
-  2: { code: "cc-by-nc", short: "CC BY-NC" },
-  3: { code: "cc-by-nc-nd", short: "CC BY-NC-ND" },
-  4: { code: "cc-by", short: "CC BY" },
-  5: { code: "cc-by-sa", short: "CC BY-SA" },
-  6: { code: "cc-by-nd", short: "CC BY-ND" },
-  7: { code: "pd", short: "PD" },
-  8: { code: "gfdl", short: "GFDL" },
-  9: { code: "cc0", short: "CC0" }
-};
+const ObservationPreload = require( "./observation_preload" );
 
 const Observation = class Observation extends Model {
   constructor( attrs, options ) {
@@ -152,199 +137,6 @@ const Observation = class Observation extends Model {
     );
   }
 
-  static preloadProjectMembership( obs, options, callback ) {
-    if ( !options.userSession || !options.userSession.user_id ) { return void callback( null ); }
-    const projobs = _.compact( _.flattenDeep( _.map( obs, o => ( o.project_observations ) ) ) );
-    const nonTradProjects = _.compact( _.flattenDeep( _.map( obs,
-      o => ( o.non_traditional_projects ) ) ) );
-    const projobsProjectIDs = _.map( projobs, po => po.project.id );
-    const nonTradProjectIDs = _.map( nonTradProjects, p => p.project_id );
-    const projectIDs = _.compact( projobsProjectIDs.concat( nonTradProjectIDs ) );
-    if ( projectIDs.length === 0 ) { return void callback( null ); }
-    const query = squel.select( ).fields( ["pu.project_id"] )
-      .from( "project_users pu" )
-      .where( "pu.project_id IN ?", projectIDs )
-      .where( "pu.user_id = ?", options.userSession.user_id );
-    pgClient.connection.query( query.toString( ), ( err, result ) => {
-      if ( err ) { return void callback( err ); }
-      const memberProjects = { };
-      _.each( result.rows, r => {
-        memberProjects[r.project_id] = true;
-      } );
-      _.each( projobs, po => {
-        po.current_user_is_member = memberProjects[po.project.id] || false;
-      } );
-      _.each( nonTradProjects, p => {
-        p.current_user_is_member = memberProjects[p.project_id] || false;
-      } );
-      callback( null );
-    } );
-  }
-
-  static preloadObservationUserPreferences( obs, callback ) {
-    if ( obs.length === 0 ) { return void callback( null ); }
-    const userIDs = _.map( obs, o => ( o.user.id ) );
-    const query = squel.select( ).fields( ["u.id, p.name, p.value"] )
-      .from( "users u" )
-      .left_join( "preferences p", null,
-        "u.id = p.owner_id AND p.owner_type= 'User' AND p.name = 'community_taxa'" )
-      .where( "u.id IN ?", userIDs );
-    pgClient.connection.query( query.toString( ), ( err, result ) => {
-      if ( err ) { return void callback( err ); }
-      const userPrefs = { };
-      _.each( result.rows, r => {
-        if ( r.value === "t" ) {
-          userPrefs[r.id] = true;
-        } else if ( r.value === "f" ) {
-          userPrefs[r.id] = false;
-        }
-      } );
-      _.each( obs, o => {
-        o.user.preferences = { prefers_community_taxa: userPrefs[o.user.id] };
-      } );
-      callback( null );
-    } );
-  }
-
-  static preloadObservationApplications( obs, callback ) {
-    if ( obs.length === 0 ) { return void callback( null ); }
-    const applicationIDs = _.uniq( _.compact( _.map( obs, o => ( o.oauth_application_id ) ) ) );
-    if ( _.isEmpty( applicationIDs ) ) { return void callback( null ); }
-    const query = squel.select( ).fields( ["id, name, url"] )
-      .from( "oauth_applications" )
-      .where( "id IN ?", applicationIDs );
-    pgClient.connection.query( query.toString( ), ( err, result ) => {
-      if ( err ) { return void callback( err ); }
-      const obsApplications = { };
-      _.each( result.rows, r => {
-        r.url = r.url || `https://www.inaturalist.org/oauth/applications/${r.id}`;
-        const parsedURL = URL.parse( r.url );
-        r.icon = `https://www.google.com/s2/favicons?domain=${parsedURL.host}`;
-        obsApplications[r.id] = r;
-      } );
-      _.each( obs, o => {
-        o.application = obsApplications[o.oauth_application_id];
-      } );
-      callback( null );
-    } );
-  }
-
-  static photoAttribution( photo, attributionName, licenseID ) {
-    const license = LICENSE_INFO[licenseID] || LICENSE_INFO[0];
-    if ( license.code === "pd" ) {
-      return `${attributionName}, no known copyright restrictions (public domain)`;
-    }
-    if ( license.code === "cc0" ) {
-      return "no rights reserved";
-    }
-    if ( license.code && license.code.match( /^cc-/ ) ) {
-      return `(c) ${attributionName}, some rights reserved (${license.short})`;
-    }
-    return `(c) ${attributionName}, all rights reserved`;
-  }
-
-  static preloadObservationPhotos( obs, callback ) {
-    if ( obs.length === 0 ) { return void callback( null ); }
-    const obsIDs = _.map( obs, "id" );
-    if ( _.isEmpty( obsIDs ) ) { return void callback( null ); }
-    const query = squel.select( ).fields( [
-      "observation_photos.id, position, uuid, photo_id, observation_id, "
-        + "photos.native_realname, photos.native_username, photos.license, "
-        + "photos.square_url, users.login, users.name"] )
-      .from( "observation_photos" )
-      .join( "photos", null, "photos.id = observation_photos.photo_id" )
-      .left_join( "users", null, "photos.user_id = users.id" )
-      .where( "observation_id IN ?", obsIDs );
-    pgClient.connection.query( query.toString( ), ( err, result ) => {
-      if ( err ) { return void callback( err ); }
-      const obsPhotos = { };
-      _.each( result.rows, r => {
-        obsPhotos[r.observation_id] = obsPhotos[r.observation_id] || [];
-        const photo = {
-          id: r.photo_id,
-          license_code: r.license === 0 ? null : LICENSE_INFO[r.license].code,
-          url: util.fix_https( r.square_url )
-        };
-        const photoAttributionName = r.native_realname || r.native_username || r.name || r.login;
-        photo.attribution = Observation.photoAttribution( photo, photoAttributionName, r.license );
-        obsPhotos[r.observation_id].push( {
-          id: r.id,
-          position: r.position,
-          uuid: r.uuid,
-          photo
-        } );
-      } );
-      const photos = _.map( _.flatten( _.values( obsPhotos ) ), "photo" );
-      if ( photos.length === 0 ) {
-        return void callback( null );
-      }
-      Model.photoDimensionsByID( photos, err2 => {
-        if ( err2 ) { return void callback( err2 ); }
-        _.each( obs, o => {
-          if ( _.isEmpty( obsPhotos[o.id] ) ) {
-            o.observation_photos = [];
-          } else {
-            o.observation_photos = _.sortBy( obsPhotos[o.id], ["position", "id"] );
-            _.each( o.observation_photos, ( op, i ) => { op.position = i; } );
-            o.photos = _.map( o.observation_photos, "photo" );
-          }
-        } );
-        const flagsQuery = squel.select( ).fields( [
-          "flag, updated_at, user_id, resolver_id, created_at, comment, id, resolved, flaggable_id"] )
-          .from( "flags" )
-          .where( "flaggable_type='Photo' AND flaggable_id IN ?", _.map( photos, "id" ) );
-        pgClient.connection.query( flagsQuery.toString( ), ( err3, flagsResult ) => {
-          if ( err3 ) { return void callback( err3 ); }
-          const photoFlags = {};
-          _.each( flagsResult.rows, r => {
-            photoFlags[r.flaggable_id] = photoFlags[r.flaggable_id] || [];
-            photoFlags[r.flaggable_id].push( r );
-          } );
-          _.each( photos, p => {
-            p.flags = photoFlags[p.id] || [];
-          } );
-          callback( null );
-        } );
-      } );
-    } );
-  }
-
-  static preloadProjectObservations( obs, callback ) {
-    if ( obs.length === 0 ) { return void callback( null ); }
-    const obsIDs = _.map( obs, "id" );
-    if ( _.isEmpty( obsIDs ) ) { return void callback( null ); }
-    const query = squel.select( ).fields(
-      ["project_observations.id, project_observations,observation_id, uuid, project_id, user_id, preferences.value"]
-    )
-      .from( "project_observations" )
-      .left_join( "preferences", null, "project_observations.id = preferences.owner_id "
-        + "AND preferences.owner_type='ProjectObservation' AND preferences.name='curator_coordinate_access'" )
-      .where( "observation_id IN ?", obsIDs );
-    pgClient.connection.query( query.toString( ), ( err, result ) => {
-      if ( err ) { return void callback( err ); }
-      const projectObs = { };
-      _.each( result.rows, r => {
-        projectObs[r.observation_id] = projectObs[r.observation_id] || [];
-        projectObs[r.observation_id].push( {
-          id: r.id,
-          uuid: r.uuid,
-          project: {
-            id: r.project_id
-          },
-          user_id: r.user_id,
-          preferences: r.value ? {
-            allows_curator_coordinate_access: r.value === "t"
-          } : {}
-        } );
-      } );
-      if ( err ) { return void callback( err ); }
-      _.each( obs, o => {
-        o.project_observations = projectObs[o.id] || [];
-      } );
-      callback( null );
-    } );
-  }
-
   static preloadAllAssociations( obs, localeOpts, callback ) {
     Observation.preloadAnnotationControlledTerms( obs, err => {
       if ( err ) { return void callback( err ); }
@@ -370,25 +162,22 @@ const Observation = class Observation extends Model {
                 po.project && po.project.project_type !== "collection" && po.project_type !== "umbrella"
               ) );
             } );
-            Observation.preloadProjectMembership( obs, localeOpts, err5 => {
-              if ( err5 ) { return void callback( err5 ); }
-              Observation.preloadObservationUserPreferences( obs, err6 => {
-                if ( err6 ) { return void callback( err6 ); }
-                Observation.preloadObservationApplications( obs, err7 => {
-                  if ( err7 ) { return void callback( err7 ); }
-                  const wikiOpts = Object.assign( { }, localeOpts, { details: true } );
-                  Taxon.assignWikipediaSummary( _.map( obs, "taxon" ), wikiOpts, err8 => {
-                    if ( err8 ) { return void callback( err8 ); }
-                    const taxonOpts = {
-                      foreignKey: "community_taxon_id",
-                      attrName: "community_taxon",
-                      modifier: t => t.prepareForResponse( localeOpts )
-                    };
-                    ESModel.fetchBelongsTo( obs, Taxon, taxonOpts, callback );
-                  } );
+            ObservationPreload.projectMembership( obs, localeOpts )
+              .then( ( ) => ObservationPreload.userPreferences( obs ) )
+              .then( ( ) => ObservationPreload.applications( obs ) )
+              .then( ( ) => {
+                const wikiOpts = Object.assign( { }, localeOpts, { details: true } );
+                Taxon.assignWikipediaSummary( _.map( obs, "taxon" ), wikiOpts, err8 => {
+                  if ( err8 ) { return void callback( err8 ); }
+                  const taxonOpts = {
+                    foreignKey: "community_taxon_id",
+                    attrName: "community_taxon",
+                    modifier: t => t.prepareForResponse( localeOpts )
+                  };
+                  ESModel.fetchBelongsTo( obs, Taxon, taxonOpts, callback );
                 } );
-              } );
-            } );
+              } )
+              .catch( e => callback( e ) );
           } );
         } );
       } );
@@ -409,8 +198,7 @@ const Observation = class Observation extends Model {
     };
     ESModel.fetchHasMany( obs, Identification, "observation.id", { modelOptions: { forObs: true } }, err => {
       if ( err ) { return void callback( err ); }
-      Observation.preloadProjectObservations( obs, err2 => {
-        if ( err2 ) { return void callback( err2 ); }
+      ObservationPreload.projectObservations( obs ).then( ( ) => {
         const ids = _.flattenDeep( _.map( obs, "identifications" ) );
         const comments = _.flattenDeep( _.map( obs, "comments" ) );
         const ofvs = _.flattenDeep( _.map( obs, "ofvs" ) );
@@ -446,11 +234,14 @@ const Observation = class Observation extends Model {
             if ( err4 ) { return void callback( err4 ); }
             ESModel.fetchBelongsTo( withUsers, User, { }, err5 => {
               if ( err5 ) { return void callback( err5 ); }
-              Observation.preloadObservationPhotos( obs, callback );
+              ObservationPreload.observationPhotos( obs )
+                .then( ( ) => callback( ) )
+                .catch( e => callback( e ) );
             } );
           } );
         } );
-      } );
+      } )
+        .catch( e => callback( e ) );
     } );
   }
 };
