+++ lib/util.js
@@ -265,6 +265,20 @@ const util = class util {
       return components;
     }, [] ).join( "&" );
   }
+
+  static photoAttribution( photo, attributionName, licenseID ) {
+    const license = util.licenseInfo[licenseID] || util.licenseInfo[0];
+    if ( license.code === "pd" ) {
+      return `${attributionName}, no known copyright restrictions (public domain)`;
+    }
+    if ( license.code === "cc0" ) {
+      return "no rights reserved";
+    }
+    if ( license.code && license.code.match( /^cc-/ ) ) {
+      return `(c) ${attributionName}, some rights reserved (${license.short})`;
+    }
+    return `(c) ${attributionName}, all rights reserved`;
+  }
 };
 
 util.iucnValues = {
@@ -277,6 +291,19 @@ util.iucnValues = {
   cr: 50,
   ew: 60,
   ex: 70
+};
+
+util.licenseInfo = {
+  0: { code: "c", short: "(c)" },
+  1: { code: "cc-by-nc-sa", short: "CC BY-NC-SA" },
+  2: { code: "cc-by-nc", short: "CC BY-NC" },
+  3: { code: "cc-by-nc-nd", short: "CC BY-NC-ND" },
+  4: { code: "cc-by", short: "CC BY" },
+  5: { code: "cc-by-sa", short: "CC BY-SA" },
+  6: { code: "cc-by-nd", short: "CC BY-ND" },
+  7: { code: "pd", short: "PD" },
+  8: { code: "gfdl", short: "GFDL" },
+  9: { code: "cc0", short: "CC0" }
 };
 
 // when util is required, it will modify the node ENV with
